requirements.txt
pafy==0.5.5
prompt-toolkit==3.0.30
python-vlc==3.0.16120
requests==2.28.1
xmltodict==0.13.0
youtube-dl==2020.12.2
youtube-search==2.1.1
